Refactored serializers and views to multiple files and moved Laboratory model to the accounts app
