+++ accounts/models/__init__.py
@@ -1,0 +1,3 @@
+from accounts.models.user import User
+from accounts.models.laboratory import Laboratory
+from accounts.models.profile import Profile
