+++ accounts/serializers/__init__.py
@@ -1,0 +1,3 @@
+from accounts.serializers.group import GroupSerializer
+from accounts.serializers.profile import ProfileSerializer
+from accounts.serializers.user import UserSerializer
