+++ accounts/serializers/group.py
@@ -1,0 +1,10 @@
+from django.contrib.auth.models import Group
+from rest_framework import serializers
+
+
+class GroupSerializer(serializers.HyperlinkedModelSerializer):
+    url = serializers.HyperlinkedIdentityField(view_name="accounts:group-detail")
+
+    class Meta:
+        model = Group
+        fields = ("name",)
