+++ accounts/views/__init__.py
@@ -1,0 +1,4 @@
+from accounts.views.group import GroupViewSet
+from accounts.views.laboratory import LaboratoryViewSet
+from accounts.views.profile import ProfileViewSet
+from accounts.views.user import UserViewSet
