+++ research/models/__init__.py
@@ -1,4 +1,3 @@
 from research.models.group import Group
-from research.models.laboratory import Laboratory
 from research.models.study import Study
 from research.models.subject import Subject
