+++ research/serializers/__init__.py
@@ -1,0 +1,3 @@
+from research.serializers.group import GroupSerializer
+from research.serializers.study import StudySerializer
+from research.serializers.subject import SubjectSerializer
