+++ research/urls.py
@@ -4,9 +4,9 @@ from rest_framework import routers
 
 app_name = "research"
 router = routers.DefaultRouter()
-router.register(r"studies", views.StudyViewSet)
-router.register(r"subjects", views.SubjectViewSet)
-router.register(r"groups", views.GroupViewSet)
+router.register(r"study", views.StudyViewSet)
+router.register(r"subject", views.SubjectViewSet)
+router.register(r"group", views.GroupViewSet)
 
 urlpatterns = [
     path("research/", include(router.urls)),
