+++ research/views/__init__.py
@@ -1,0 +1,3 @@
+from research.views.group import GroupViewSet
+from research.views.study import StudyViewSet
+from research.views.subject import SubjectViewSet
